- Merged relay server.
- Fixes some name consistency.
- Fixed multiple crashes.
- Fixed DNA bug.

+++ src/Projectile.cs
@@ -35,9 +35,7 @@ public class Projectile : Actor {
 	public bool shouldVortexSuck = true;
 	bool damagedOnce;
 	public ShaderWrapper nightmareZeroShader;
-	public bool useTransparentFade;
 	//public int? destroyFrames;
-
 	public Player ownerPlayer;
 
 	public Projectile(Weapon weapon, Point pos, int xDir, float speed, float damage, Player player, string sprite, int flinch, float hitCooldown, ushort? netId, bool ownedByLocalPlayer) : base(sprite, pos, netId, ownedByLocalPlayer, false) {
@@ -91,14 +89,8 @@ public class Projectile : Actor {
 		*/
 
 		if (time > maxTime || moveDistance > maxDistance || pos.x > Global.level.width + leeway || pos.x < -leeway || pos.y > Global.level.height + leeway || pos.y < -leeway) {
-			if (fadeOnAutoDestroy) {
-				destroySelf();
-			} else {
-				if (useTransparentFade) {
-					new ProjFadeAnim(pos, vel, sprite.name, sprite.frameIndex, xDir, zIndex, null, true);
-				}
-				destroySelfNoEffect();
-			}
+			if (fadeOnAutoDestroy) destroySelf();
+			else destroySelfNoEffect();
 			return;
 		}
 	}
